add ext_todo-20030105.patch
---When email comes into the server extremely quickly, qmail can
sometimes spend so much time pre-processing the email that the
email starts backing up and most of it is in an undeliverable
state in the todo queue rather than the real queue (note that
this is only a problem when you get massive amounts of email
at the same time (several per second)). This behavior is
frequently referred to as "silly qmail syndrome." André Opperman
wrote a patch to solve the problem. It makes qmail-todo a
separate program, allowing the pre-processing to happen
asynchronously and thus does not prevent deliveries from ocurring
at the same time. The solution is a bit complicated, and is not
worth applying unless you are experiencing "silly qmail syndrome."
This patch is referred to as the ext_todo patch.---